Make all tests work
