+++ tests/skipLinksTestDataset/config_template.ini
@@ -92,10 +92,10 @@ h5-event-out-dir = {{ embryonicDir }}/tests/skipLinksTestDataset/events
 #ctc-filename-zero-pad-length = 3
 #ctc-output-dir = {{ embryonicDir }}/tests/skipLinksTestDataset/ctc_RES
 
-graph-json-file = {{ embryonicDir }}/tests/skipLinksTestDataset/out-graph.json
-result-json-file = {{ embryonicDir }}/tests/skipLinksTestDataset/out-result.json
-label-image-file = {{ embryonicDir }}/tests/skipLinksTestDataset/out-label-image.h5
-label-image-path = /TrackingFeatureExtraction/LabelImage/0000/[[%d, 0, 0, 0, 0], [%d, %d, %d, %d, 1]]
+#graph-json-file = {{ embryonicDir }}/tests/skipLinksTestDataset/out-graph.json
+#result-json-file = {{ embryonicDir }}/tests/skipLinksTestDataset/out-result.json
+#label-image-file = {{ embryonicDir }}/tests/skipLinksTestDataset/out-label-image.h5
+#label-image-path = /TrackingFeatureExtraction/LabelImage/0000/[[%d, 0, 0, 0, 0], [%d, %d, %d, %d, 1]]
 
 
 ctc-output-dir = {{ embryonicDir }}/tests/skipLinksTestDataset/ctc_RES
