+++ tests/test_pipeline.py
@@ -109,9 +109,35 @@ def test_noscripts_mergerResolvingTestDataset():
     with open('../tests/mergerResolvingTestDataset/ctc_RES/res_track.txt', 'r') as resultFile:
         assert(resultFile.read() == '1 0 3 0\n2 0 3 0\n')
 
+def test_skipLinksTestDataset_withoutTracklets():
+    check_call(["python",
+                "../hytra/configtemplates/create_config.py",
+                "--in",
+                "../tests/skipLinksTestDataset/config_template.ini",
+                "--out",
+                "../tests/skipLinksTestDataset/test_config.ini",
+                "embryonicDir",
+                "..",
+                ])
+    check_call(["python", "pipeline_skip_links.py", "--config", "../tests/skipLinksTestDataset/test_config.ini"])
+
+    for f in range(4):
+        frame = vigra.impex.readImage('../tests/skipLinksTestDataset/ctc_RES/mask00{}.tif'.format(f))
+        if f == 1: # the second frame is empty
+            assert(np.all(np.unique(frame) == [0]))
+        elif f==0:
+            assert(np.all(np.unique(frame) == [0, 1]))
+        else:
+            assert(np.all(np.unique(frame) == [0, 2]))
+
+
+    with open('../tests/skipLinksTestDataset/ctc_RES/res_track.txt', 'r') as resultFile:
+        assert(resultFile.read() == '1 0 0 0\n2 2 3 1\n')
+
 if __name__ == "__main__":
     test_divisionTestDataset()
     test_mergerResolvingTestDataset()
     test_divisionTestDataset_withoutTracklets()
     test_mergerResolvingTestDataset_withoutTracklets()
     test_noscripts_mergerResolvingTestDataset()
+    test_skipLinksTestDataset_withoutTracklets()
